feat(tickets): create TicketContent component for improved ticket and payment navigation
